// frontend/src/packages/__generated__/PackageSearchFragment_packages.graphql.ts
export type PackageSearchFragment_packages$data = {
  readonly packages: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly " $fragmentSpreads": FragmentRefs<"PackageSearchListItemFragment_package">;
      } | null | undefined;
    } | null | undefined> | null | undefined;
    readonly totalCount: number;
  };
  readonly " $fragmentType": "PackageSearchFragment_packages";
};
export type PackageSearchFragment_packages$key = {
  readonly " $data"?: PackageSearchFragment_packages$data;
  readonly " $fragmentSpreads": FragmentRefs<"PackageSearchFragment_packages">;
};